// src/app/core/organizer/enums.ts
export enum FolderEditMode {
    Create = 'CREATE',
    Rename = 'RENAME',
    Delete = 'DELETE',
}

export enum CalendarGroupEditMode {
    Create = 'CREATE',
    Rename = 'RENAME',
    Delete = 'DELETE',
    CreateCalendar = 'CREATECALENDAR',
}

export enum CalendarEditMode {
    Rename = 'RENAME',
    Delete = 'DELETE',
    Color = 'COLOR',
    NewBitrhday = 'NEWBIRTHDAY'
}

export enum FolderEditKind {
    Init = 'INIT',
    Confirm = 'CONFIRM',
    Reject = 'REJECT'
}